Updated to use the DOM rather than mongo.
The template wasn't re-rendering because the data never changes; so
use the DOM instead! I also renamed everything because I had to create
a new project in order to make this work. The single stock template
could very easily not exist, but it's late and I didn't feel like
making that change.